Resolve PR comments

+++ db/migrate/20201001134619_remove_not_null_constraint_from_mapping_terms_predicate_field.rb
@@ -2,6 +2,6 @@
 
 class RemoveNotNullConstraintFromMappingTermsPredicateField < ActiveRecord::Migration[6.0]
   def change
-    change_column :mapping_terms, :predicate_id, :integer, null: true
+    change_column :mapping_terms, :predicate_id, :bigint, null: true
   end
 end
